# docs/swagger.yaml
basePath: /
definitions:
  requestmodels.ContactInfo:
    properties:
      Mail:
        type: string
      Phone:
        type: string
    type: object
...
  requestmodels.TenantInfo:
    properties:
      CriminalRecord:
        type: boolean
      Income:
        type: number
      IncomeProof:
        type: boolean
    type: object
  requestmodels.User:
    properties:
      Apartment:
        type: string
      City:
        type: string
      Country:
        type: string
      First:
        type: string
      Housenumber:
        type: string
      Id:
        type: string
      Last:
        type: string
      Mail:
        type: string
      Phone:
        type: string
      Street:
        type: string
    type: object
...
  responsemodels.TenantInfo:
    properties:
      CriminalRecord:
        type: boolean
      Id:
        type: string
      Income:
        type: number
      IncomeProof:
        type: boolean
    type: object
  responsemodels.User:
    properties:
      Apartment:
        type: string
      City:
        type: string
      Country:
        type: string
      First:
        type: string
      Housenumber:
        type: string
      Id:
        type: string
      Last:
        type: string
      Mail:
        type: string
      Phone:
        type: string
      Street:
        type: string
    type: object
host: localhost:3000
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Fiber Swagger Example API
  version: "2.0"
paths:
...
  /user/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Delete a user
      tags:
      - user
    get:
      consumes:
      - application/json
      description: Get a user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "302":
          description: Found
          schema:
            $ref: '#/definitions/responsemodels.User'
      summary: Get a user
      tags:
      - user
  /user/contactinfo/{id}:
    put:
      consumes:
      - application/json
      description: Update a user's contact info
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Contact Info
        in: body
        name: contactInfo
        required: true
        schema:
          $ref: '#/definitions/requestmodels.ContactInfo'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responsemodels.User'
      summary: Update a user's contact info
      tags:
      - user
  /user/create:
    post:
      consumes:
      - application/json
      description: Create a new user
      parameters:
      - description: User
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/requestmodels.User'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/responsemodels.User'
      summary: Create a new user
      tags:
      - user
...
  /user/tenantinfo/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a user's tenant info
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Delete a user's tenant info
      tags:
      - user
    get:
      consumes:
      - application/json
      description: Get a user's tenant info
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "302":
          description: Found
          schema:
            $ref: '#/definitions/responsemodels.TenantInfo'
      summary: Get a user's tenant info
      tags:
      - user
    post:
      consumes:
      - application/json
      description: Create a user's tenant info
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Tenant Info
        in: body
        name: tenantInfo
        required: true
        schema:
          $ref: '#/definitions/requestmodels.TenantInfo'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/responsemodels.TenantInfo'
      summary: Create a user's tenant info
      tags:
      - user
    put:
      consumes:
      - application/json
      description: Update a user's tenant info
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Tenant Info
        in: body
        name: tenantInfo
        required: true
        schema:
          $ref: '#/definitions/requestmodels.TenantInfo'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responsemodels.TenantInfo'
      summary: Update a user's tenant info
      tags:
      - user
  /user/update:
    put:
      consumes:
      - application/json
      description: Update a user
      parameters:
      - description: User
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/requestmodels.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responsemodels.User'
      summary: Update a user
      tags:
      - user
schemes:
- http
swagger: "2.0"